Fixed bugs and added benchmarking

+++ hashset.h
@@ -6,22 +6,22 @@
 
 struct Node {
     void* key;
-    int hash_value;
+    size_t hash_value;
     Node* next;
 
     Node() : key(nullptr), hash_value(0), next(nullptr) {}
     Node(void* ptr, int hv, Node* next) : key(ptr), hash_value(hv), next(next) {}
 
     void* getKey() { return key; }
-    int getHashValue() { return hash_value; }
+    size_t getHashValue() { return hash_value; }
     Node* getNext() { return next; }
 };
 
 uint64_t get_murmur_hashcode(void* ptr);
 
 struct HashSet {
-    int size;
-    int occupiedSlots;
+    size_t size;
+    size_t occupiedSlots;
     Node** items;
 
     HashSet();
@@ -34,7 +34,7 @@ struct HashSet {
 private:
     bool isEligibleForRehashForInsert();
     bool isEligibleForRehashForRemove();
-    void rehash(int newLen);
+    void rehash(size_t newLen);
 };
 
 #endif // HASHSET_H
